+++ share/objects/specials/malachite/02-vivid-malachite.lua
@@ -1,0 +1,8 @@
+local map = ...
+local objects = map.objects
+local globals = map.globals
+
+local id = require ('lib.gem.id')
+
+-- Vivid Malachite
+local unit = objects [id (globals.Gem_Special__MALACHITE_2)]